Add new `system status` command
Container services can be started and stopped using the `system start`
and `system stop` commands. But there is no straightforward way for
users to check if the services were already started.

This commit adds a new `system status` command that shows wether the
container services are running or not.

The command only checks the status of the apiserver for now, but it
could be expanded in the future to show more details.