Bump Containerization to 0.5.0
I'm happy with the config improvements, but because
IO can only be setup at constructor time this makes it
so that we need to supply IO at creation time of the VM
or exec, which isn't the end of the world. So now boostrap()
and createProcess() take in the filehandles.