Bump Containerization to 0.5.0
I'm happy with the config improvements, but because
IO can only be setup at constructor time this makes it
quite awkward. I see two ways to make it less weird:

1. Supply IO at bootstrap time.
2. Move to a model of IO being returned from the daemon.

This change additionally adds in a boolean to ContainerConfiguration
to ask to open stdin. The way we were doing this was by checking if
the 0th index in the filehandles we receive is non-nil, but this
is a bit crummy for -i -d.