CLI (run): Only ask for stdin for -i
Today if you asked for a pty and also wanted to detach we'd open stdin and then
immediately close that pipe which would close the guest relay as well. I don't
believe we need stdin open unless it's asked for with -i. Truthfully there's
one extra bit here that is needed which is if -i and -d are supplied we need to
tell the daemon to open stdin, but not send an fd (as they're client supplied and
the client/cli is going to exit immediately). That will need to be a followup, as
it's mainly useful for attach which we don't have today.