APIServer: Make container deletes a little more robust
Today it's possible we receive the container exit event sent to the
daemon after a user might've invoked `container delete`. The aftermath
of that unfortunate scenario is that you would be met with an error stating
we can't delete the container because it is running. This is not really common
unless someone was doing kill/stop+delete in some script, but it still is an issue
that needs to be fixed. The true source of truth for state is actually the sandboxservice
hosting the container, so if we were guaranteed that that is still alive, we can
query and take some action based on the possible mismatch in current snapshot state
and what the sb service is telling us.

To make this possible I've introduced a new rpc on the sandbox service named shutdown.
Shutdown's purpose is to run some cleanup code (if there is any) and then to wind down
the process. Prior to this RPC the sandbox service would basically just call exit(0)
on its own accord after the container exited. Now this process is driven by the APIServer
instead, which I like a little bit more. The true win here though is that because we know
as long as we haven't sent the shutdown RPC that the process is alive (barring a bug), we
can query the state of the container the whole time.