Make container start idempotent
Today it fails in bootstrap the second go around, and we also
have an error handler that automatically cleans up the container
if bootstrap failed which is even worse. This short circuits us
first in the cli if the state is running when we get() the container,
and also adds in a clause to bootstrap to just early return if we
already have a client.