Implement container stats
This implements statistics gathering across the various components, but
ultimately this is for implementing a new CLI command: `container stats`.
This shows memory usage, cpu usage, network and block i/o and the number
of processes in the container. The new command can inspect stats for 1-N
containers and by default continuously updates in a `top` like stream.