fix(TerminalProgress): make the progress bar respect locale-specific decimal separator
The `ProgressBar#adjustFormattedSize` function currently expects a
decimal dot when adding the additional ".0" to the size. This, however,
breaks when a region with a non-dot decimal separator is used.

This commit makes the function use the decimal separator of the current
locale instead, making it work as intended under environments that don't
use the decimal dot.