CLI: Fix -it not being able to pipe stdout
Typically if one fd is a tty, it's common for all 3 of stdio to be the same,
but that is not always the case. In our case we were using our Terminal type
from Containerization to comb through err/out/in and give us a type backed by
one of the 3 if -t was supplied. It happens that stderr is the first we check,
so our Terminal() is backed by fd 2. This change modifies things so that we
always initialize our Terminal if asked for with fd 0, and out/err are backed
by their corresponding correct fd number.