CLI: Fix stop not signalling waiters
I don't believe this is the actual fix for some of the hangs we've been seeing, but this is definitely
wrong. For the Sandbox Svcs stop implementation, it did not signal the wait continuations to unblock any
pending wait rpcs. What you'd typically get if you had a running container (that wasn't --detach) is the
dreaded xpc interrupted error for the wait rpc. After this change they should be resumed.